+++ src/main/java/com/example/pingpong/user/controller/UserController.java
@@ -1,11 +1,15 @@
-package com.example.pingpong.controller;
+package com.example.pingpong.user.controller;
 
-import com.example.pingpong.model.User;
-import com.example.pingpong.service.UserService;
+import com.example.pingpong.user.model.User;
+import com.example.pingpong.user.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Map;
 
 @RestController
 @AllArgsConstructor
@@ -21,11 +25,20 @@ public class UserController {
     }
 
     @PostMapping("/guest-login")
-    public ResponseEntity<User> createGuestAccount() {
+    public ResponseEntity<User> createGuestAccount(HttpServletRequest request) {
         User guestUser = userService.createGuestAccount();
+
+        request.getSession().setAttribute("user",guestUser);
         return ResponseEntity.ok(guestUser);
     }
 
+
+    @RequestMapping("/test")
+    public void test(HttpServletRequest request, HttpServletResponse response) {
+
+        Map<String, Object> user = (Map<String, Object>) request.getAttribute("user");
+        String nickname = user.get("nickname").toString();
+    }
     @GetMapping("/{nickName}")
     public ResponseEntity<User> getUserById(@PathVariable String nickName) {
         return userService.findByNickname(nickName)
