+++ src/main/java/com/example/pingpong/user/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.example.pingpong.exception;
+package com.example.pingpong.user.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
