+++ src/main/java/com/example/pingpong/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.pingpong.repository;
+package com.example.pingpong.user.repository;
 
-import com.example.pingpong.model.User;
+import com.example.pingpong.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
