+++ src/main/resources/application.yml
@@ -7,7 +7,7 @@ spring:
   jpa:
     show-sql: true
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
       format_sql: true
     properties:
       hibernate:
